Add an Api class for fetch calls and cleaners, refactor App
