--- src/FormContainer/FormContainer.css
@@ -1,75 +1,0 @@
-.tab-container {
-  width: 80%;
-  margin: 45px auto;
-  display: flex;
-  flex-wrap: wrap;
-  height: 515px;
-}
-
-.radio-btn {
-  display: none;
-}
-
-.tab-label {
-  order: 1;
-  display: block;
-  width: 25%;
-  height: 50px;
-  color: #757575;
-  text-align: center;
-  background: #f0f0f0;
-  border-bottom: 1px solid grey;
-  padding: 15px;
-}
-
-.content {
-  order: 99;
-  flex-grow: 1;
-  width: 100%;
-  display: none;
-  border: 1px solid grey;
-  border-top: none;
-  padding: 40px 25px;
-  /*height: auto;*/
-  max-height: 375px;
-  overflow-y: scroll;
-}
-
-.tab-container input[type='radio'] {
-  position: absolute;
-}
-
-.tab-container input[type='radio']:checked + label  {
-  background: #fff;
-  border: 1px solid grey;
-  border-bottom: none;
-  font-weight: 600;
-}
-
-.tab-container input[type='radio']:checked + label + .content {
-  display: unset;
-/*  grid-template-columns: 35% 65%;
-  grid-template-rows: auto auto;*/
-}
-
-.box {
-  grid-column: 1;
-  grid-row: 1 / 3;
-  width: 90%;
-  height: 100%;
-  border: 1px solid grey;
-  /*background-image: url('./placeholder.png');*/
-  background-size: 100% 100%;
-}
-
-.content h2 {
-  grid-column: 2;
-  grid-row: 1;
-  border-bottom: 1px solid grey;
-}
-
-.content p {
-  grid-column: 2;
-  grid-row: 2;
-  padding-top: 20px;
-}
